Address Copilot review: allowlist columns, fix flaky test, extract helpers
- Add UPDATABLE_COLUMNS allowlist to updatePet() so only known pet
  columns are interpolated into SQL, preventing unknown-column errors
  and closing the injection surface.
- Fix flaky delete test: use Playwright's auto-retrying `toHaveCount`
  instead of reading `.pet-count` text synchronously.
- Extract shared E2E helpers (waitForAppReady, waitForPets, etc.) into
  tests/e2e/helpers.js to eliminate duplication between spec files.
