+++ src/lib/services/petService.ts
@@ -214,6 +214,23 @@ export async function uploadPet(
   };
 }
 
+/** Columns that callers are allowed to update via updatePet(). */
+const UPDATABLE_COLUMNS = new Set([
+  'name',
+  'gender',
+  'breed',
+  'notes',
+  'genome_data',
+  'intelligence',
+  'toughness',
+  'friendliness',
+  'ruggedness',
+  'enthusiasm',
+  'virility',
+  'ferocity',
+  'temperament',
+]);
+
 /**
  * Update a pet record.
  */
@@ -233,7 +250,7 @@ export async function updatePet(petId: number, updates: Record<string, unknown>)
   }
 
   for (const [field, value] of Object.entries(flat)) {
-    if (field === 'id') continue;
+    if (!UPDATABLE_COLUMNS.has(field)) continue;
     if (field === 'genome_data') {
       setClauses.push(`${field} = ?`);
       values.push(typeof value === 'string' ? value : JSON.stringify(value));
